- moved constants to storage context - merged dependencies into package.json of repo
