+++ package.json
@@ -9,7 +9,10 @@
     "@openzeppelin/hardhat-upgrades": "^1.21.0",
     "chai": "^4.3.7",
     "ethers": "^5.7.2",
-    "hardhat": "^2.14.0"
+    "hardhat": "^2.14.0",
+    "axios": "^1.4.0"
   },
-  "dependencies": {}
+  "dependencies": {
+    "@tenderly/actions": "^0.2.4"
+  }
 }
