--- web3-actions/actions/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compileOnSave": true,
-  "compilerOptions": {
-    "module": "commonjs",
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "outDir": "out",
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2020",
-    "resolveJsonModule": true,
-    "esModuleInterop": true
-  },
-  "include": ["./trigger.ts"]
-}
